feat: more interfaces exported + code cleanup
