+++ src/types/hip412-validator.ts
@@ -1,0 +1,34 @@
+import { MetadataObject } from './csv';
+
+export interface FileValidationResult {
+  isValid: boolean;
+  fileName?: string;
+  errors: string[];
+}
+
+export interface DetailedFileValidationResult {
+  isValid: boolean;
+  errors: string[];
+  errorsCount: number;
+}
+
+export interface ValidateArrayOfObjectsResult {
+  results: { [index: number]: DetailedFileValidationResult };
+  generalValid: boolean;
+}
+
+export interface DirectoryValidationResult {
+  isValid: boolean;
+  errors: MetadataError[];
+}
+
+export interface MetadataError {
+  fileName?: string;
+  general: string[];
+}
+
+export interface MetadataOnChainObjects {
+  metadata?: MetadataObject;
+  serialNumber: number;
+  error?: string;
+}
